+++ build.sh
@@ -59,69 +59,14 @@ set -x
 ###################
 # frontend # (output: frontend/public/)
 ###################
-(cd "$TOPLEVEL/frontend" && npm ci && npm run build)
-
-#################
-# assets.tar.xz #
-#################
-
-# we need GNU tar (see below) so we check early
-if tar --help | grep GNU >/dev/null; then
-  echo "found GNU tar as tar"
-  tar="tar"
-elif command -v gtar >/dev/null; then
-  echo "found GNU tar as gtar"
-  tar="gtar"
-else
-  echo "did not find GNU tar, please install"
-  echo "  brew install gnu-tar"
-  exit 1
-fi
-
-if ! command -v xz >/dev/null; then
-  echo "did not find xz, please install"
-  echo "  brew install xz"
-  exit 1
-fi
-
-# We use a local directory, and we don't delete it after the build, so that
-# assets can be inspected.
-tarball_dir="$TOPLEVEL/web-assets"
-rm -rf "$tarball_dir"
-echo "using $tarball_dir for tarball directory"
-cp -R "$TOPLEVEL/frontend/public/" "$tarball_dir/"
-
-# Bundle into a tight tarball
-# On macOS you need to install gtar + xz
-# brew install gnu-tar
-# brew install xz
-cd "$tarball_dir"
-
-"$tar" cJv --mtime='2021-05-07 17:00+00' --sort=name --exclude .last_build_id -f "$TOPLEVEL/assets.tar.xz" .
-
-cd "$TOPLEVEL"
-
-ls -sh "$TOPLEVEL/assets.tar.xz"
-sha256sum "$TOPLEVEL/assets.tar.xz"
+"$TOPLEVEL/build-frontend.sh"
 
 ###############
-# cargo build # (output: target/release/.../nns-dapp.wasm)
+# backend # (output: nns-dapp.wasm)
 ###############
 echo Compiling rust package
-cargo_args=(--target wasm32-unknown-unknown --release --package nns-dapp)
 if [[ $DFX_NETWORK != "mainnet" ]]; then
-  cargo_args+=(--features mock_conversion_rate)
+  "$TOPLEVEL/build-rs.sh" nns-dapp --features mock_conversion_rate
+else
+  "$TOPLEVEL/build-rs.sh" nns-dapp
 fi
-
-(cd "$TOPLEVEL" && cargo build "${cargo_args[@]}")
-
-####################
-# ic-cdk-optimizer # (output: nns-dapp.wasm)
-####################
-echo Optimising wasm
-cd "$TOPLEVEL"
-ic-cdk-optimizer ./target/wasm32-unknown-unknown/release/nns-dapp.wasm -o ./nns-dapp.wasm
-gzip -f -n nns-dapp.wasm
-mv nns-dapp.wasm.gz nns-dapp.wasm
-ls -sh ./nns-dapp.wasm
-sha256sum ./nns-dapp.wasm