+++ frontend/jest-setup.ts
@@ -36,8 +36,9 @@ jest.mock("./src/lib/constants/environment.constants.ts", () => ({
   FETCH_ROOT_KEY: false,
   ENABLE_SNS_2: true,
   ENABLE_SNS_VOTING: true,
-  ENABLE_SNS_CACHING: true,
-  CACHING_CANISTER_URL:
+  ENABLE_SNS_AGGREGATOR: true,
+  ENABLE_CKBTC_LEDGER: true,
+  SNS_AGGREGATOR_CANISTER_URL:
     "https://5v72r-4aaaa-aaaaa-aabnq-cai.small12.testnet.dfinity.network",
   STAKE_MATURITY: true,
 }));