+++ frontend/src/lib/api/sns-ledger.api.ts
@@ -1,12 +1,10 @@
+import { getIcrcMainAccount } from "$lib/api/icrc-ledger.api";
 import type { SubAccountArray } from "$lib/canisters/nns-dapp/nns-dapp.types";
 import type { Account } from "$lib/types/account";
-import { LedgerErrorKey } from "$lib/types/ledger.errors";
 import { nowInBigIntNanoSeconds } from "$lib/utils/date.utils";
 import { logWithTimestamp } from "$lib/utils/dev.utils";
-import { mapOptionalToken } from "$lib/utils/sns.utils";
 import type { Identity } from "@dfinity/agent";
-import { encodeIcrcAccount, type IcrcAccount } from "@dfinity/ledger";
-import { TokenAmount } from "@dfinity/nns";
+import type { IcrcAccount } from "@dfinity/ledger";
 import type { Principal } from "@dfinity/principal";
 import { arrayOfNumberToUint8Array, toNullable } from "@dfinity/utils";
 import { wrapper } from "./sns-wrapper.api";
@@ -22,35 +20,22 @@ export const getSnsAccounts = async ({
 }): Promise<Account[]> => {
   // TODO: Support subaccounts
   logWithTimestamp("Getting sns accounts: call...");
+
   const { balance, ledgerMetadata } = await wrapper({
     identity,
     rootCanisterId: rootCanisterId.toText(),
     certified,
   });
 
-  const snsMainAccount = { owner: identity.getPrincipal() };
-  const [mainBalanceE8s, metadata] = await Promise.all([
-    balance(snsMainAccount),
-    ledgerMetadata({}),
-  ]);
-
-  const projectToken = mapOptionalToken(metadata);
-
-  if (projectToken === undefined) {
-    throw new LedgerErrorKey("error.sns_token_load");
-  }
-
-  const mainAccount: Account = {
-    identifier: encodeIcrcAccount(snsMainAccount),
-    principal: identity.getPrincipal(),
-    balance: TokenAmount.fromE8s({
-      amount: mainBalanceE8s,
-      token: projectToken,
-    }),
-    type: "main",
-  };
+  const mainAccount = await getIcrcMainAccount({
+    identity,
+    certified,
+    balance,
+    metadata: ledgerMetadata,
+  });
 
   logWithTimestamp("Getting sns accounts: done");
+
   return [mainAccount];
 };
 
@@ -76,6 +61,14 @@ export const transactionFee = async ({
   return fee;
 };
 
+/**
+ * Transfer SNS tokens from one account to another.
+ *
+ * param.fee is mandatory to ensure that it's show for hardware wallets.
+ * Otherwise, the fee would not show in the device and the user would not know how much they are paying.
+ *
+ * This als adds an extra layer of safety because we show the fee before the user confirms the transaction.
+ */
 export const transfer = async ({
   identity,
   to,
@@ -84,6 +77,7 @@ export const transfer = async ({
   memo,
   fromSubAccount,
   createdAt,
+  fee,
 }: {
   identity: Identity;
   to: IcrcAccount;
@@ -92,6 +86,7 @@ export const transfer = async ({
   memo?: Uint8Array;
   fromSubAccount?: SubAccountArray;
   createdAt?: bigint;
+  fee: bigint;
 }): Promise<void> => {
   const { transfer: transferApi } = await wrapper({
     identity,
@@ -106,6 +101,7 @@ export const transfer = async ({
       subaccount: toNullable(to.subaccount),
     },
     created_at_time: createdAt ?? nowInBigIntNanoSeconds(),
+    fee,
     memo,
     from_subaccount:
       fromSubAccount !== undefined