+++ frontend/src/lib/services/sns-accounts.services.ts
@@ -2,6 +2,7 @@ import { getSnsAccounts, transfer } from "$lib/api/sns-ledger.api";
 import { snsAccountsStore } from "$lib/stores/sns-accounts.store";
 import { snsTransactionsStore } from "$lib/stores/sns-transactions.store";
 import { toastsError } from "$lib/stores/toasts.store";
+import { transactionsFeesStore } from "$lib/stores/transaction-fees.store";
 import type { Account } from "$lib/types/account";
 import { toToastError } from "$lib/utils/error.utils";
 import { ledgerErrorToToastError } from "$lib/utils/sns-ledger.utils";
@@ -9,6 +10,7 @@ import { numberToE8s } from "$lib/utils/token.utils";
 import type { Identity } from "@dfinity/agent";
 import { decodeIcrcAccount } from "@dfinity/ledger";
 import type { Principal } from "@dfinity/principal";
+import { get } from "svelte/store";
 import { getAuthenticatedIdentity } from "./auth.services";
 import { loadAccountTransactions } from "./sns-transactions.services";
 import { loadSnsTransactionFee } from "./transaction-fees.services";
@@ -44,7 +46,7 @@ export const loadSnsAccounts = async ({
       toastsError(
         toToastError({
           err,
-          fallbackErrorLabelKey: "error.sns_accounts_load",
+          fallbackErrorLabelKey: "error.accounts_load",
         })
       );
 
@@ -86,12 +88,20 @@ export const snsTransferTokens = async ({
     const identity: Identity = await getSnsAccountIdentity(source);
     const to = decodeIcrcAccount(destinationAddress);
 
+    const fee = get(transactionsFeesStore).projects[rootCanisterId.toText()]
+      ?.fee;
+
+    if (!fee) {
+      throw new Error("error.transaction_fee_not_found");
+    }
+
     await transfer({
       identity,
       to,
       fromSubAccount: source.subAccount,
       e8s,
       rootCanisterId,
+      fee,
     });
 
     await Promise.all([