+++ frontend/src/lib/types/i18n.d.ts
@@ -41,9 +41,9 @@ interface I18nError {
   neuron_spawning: string;
   neuron_load: string;
   sns_neurons_load: string;
-  sns_accounts_load: string;
+  accounts_load: string;
   sns_accounts_balance_load: string;
-  sns_token_load: string;
+  icrc_token_load: string;
   list_proposals: string;
   list_canisters: string;
   missing_identity: string;